requirements.txt
Flask==2.3.3
Flask-RESTX==1.3.0
PyJWT==2.8.0
cryptography==45.0.7
requests==2.32.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
minio==7.2.0
pytest==7.4.3
